core/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
numpy==2.0.2
scipy==1.14.1
scikit-learn==1.5.2
pgvector==0.3.6
sentence-transformers==3.3.1
pennylane==0.39.0
autoray==0.6.12
httpx==0.28.1
ir-measures==0.3.7
pytest==8.3.4
